image renders

+++ crates/shader-types/src/padding.rs
@@ -1,9 +1,10 @@
 use super::bytemuck::{Pod, Zeroable};
+
 #[derive(Clone, Copy, Debug, derive_more::From)]
 #[repr(C)]
 pub struct WithPadding<const N: usize, T> {
     pub inner: T,
-    padding: [u32; N],
+    padding: [f32; N],
 }
 
 impl<const N: usize, T> core::default::Default for WithPadding<N, T>
@@ -17,7 +18,7 @@ where
 
 impl<const N: usize, T> WithPadding<N, T> {
     pub const fn pad(inner: T) -> Self {
-        Self { inner, padding: [0; N] }
+        Self { inner, padding: [0.; N] }
     }
 }
